+++ apps/pico/cli/translations.ts
@@ -1,0 +1,7 @@
+import { tx } from "@use-pico/server";
+
+tx({
+	packages: [`${__dirname}/..`],
+	output: `${__dirname}/../src/translation`,
+	locales: ["cs", "en"],
+});