+++ apps/pico/src/@routes/$locale/apps.tsx
@@ -1,0 +1,11 @@
+import { createFileRoute } from "@tanstack/react-router";
+import { ls } from "@use-pico/client";
+import { SessionSchema } from "~/app/schema/SessionSchema";
+
+export const Route = createFileRoute("/$locale/apps")({
+	loader() {
+		return {
+			session: SessionSchema.parse(ls.get("session")),
+		};
+	},
+});