+++ apps/pico/src/@routes/$locale/apps/index.tsx
@@ -1,0 +1,9 @@
+import { createFileRoute, useLoaderData } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/$locale/apps/")({
+	component: () => {
+		const { session } = useLoaderData({ from: "/$locale/apps" });
+
+		return <pre>{JSON.stringify(session, null, 2)}</pre>;
+	},
+});