+++ apps/pico/src/@routes/$locale/index.tsx
@@ -1,17 +1,21 @@
-import { createFileRoute } from "@tanstack/react-router";
+import { createFileRoute, redirect } from "@tanstack/react-router";
+import { ls } from "@use-pico/client";
+import { SessionSchema } from "~/app/schema/SessionSchema";
 
 export const Route = createFileRoute("/$locale/")({
-	beforeLoad: async ({ context: { queryClient }, params: { locale } }) => {
-        //
-        
-		// if (ticket) {
-		// 	throw redirect({
-		// 		to: `/$locale/${ticket.site}`,
-		// 	});
-		// }
-		// throw redirect({
-		// 	to: `/$locale/public`,
-		// 	params: { locale },
-		// });
+	beforeLoad: async ({ params: { locale } }) => {
+		const sessionResult = SessionSchema.safeParse(ls.get("session"));
+
+		if (sessionResult.success) {
+			throw redirect({
+				to: `/$locale/apps`,
+				params: { locale },
+			});
+		}
+
+		throw redirect({
+			to: `/$locale/public/login`,
+			params: { locale },
+		});
 	},
 });