+++ apps/pico/src/app/schema/LoginSchema.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 
 export const LoginSchema = z.object({
-	login: z.string(),
-	password: z.string(),
+	login: z.string().min(1),
+	password: z.string().min(1),
 });
 
 export type LoginSchema = typeof LoginSchema;