+++ apps/pico/src/app/type/UserType.ts
@@ -1,0 +1,7 @@
+import type { IdentitySchema } from "@use-pico/common";
+
+export interface UserType extends IdentitySchema.Type {
+	name: string;
+	login: string;
+	password: string;
+}