+++ apps/pico/src/translation/dynamic.ts
@@ -1,0 +1,9 @@
+import { translator } from "@use-pico/common";
+
+/**
+ * List of dynamic translations picked up by the translation extractor.
+ *
+ * This piece of code is not used in the application.
+ */
+translator.rich("String must contain at least 1 character(s)");
+translator.rich("Passwords must match");