+++ packages/client/src/ls/ls.ts
@@ -1,0 +1,3 @@
+import lls from "localstorage-slim";
+
+export const ls = lls;