+++ packages/common/src/pwd/pwd.ts
@@ -1,0 +1,12 @@
+import { sha224 } from "js-sha256";
+
+export const pwd = {
+	hash(password: string) {
+		return sha224(password);
+	},
+	verify(password: string, hash: string) {
+		return sha224(password) === hash;
+	},
+} as const;
+
+export type pwd = typeof pwd;